Assistant checkpoint: Fix image thumbnail and delete button positioning
Assistant generated file changes:
- client/src/components/ui/image-carousel.tsx: Fix image thumbnail and delete button sizing/positioning

---

User prompt:

You are making random code change without using critical thhinking and it frustrates me. This is the 3rd time I am telling you that the images that we upload are showing as 20x20 thumbnails. The delete icon is set for 40x20 thumbnail size. We need to make sure the delete button is repositioned to our 20x20 thumbnails